+++ client/NetworkClient.h
@@ -33,8 +33,8 @@ public:
 
 	// Commands
 	virtual void connect_command() = 0;
-	virtual void move_command(int dir) = 0;
-	virtual void turn_command(int dir) = 0;
+	virtual void move_command(double dir) = 0;
+	virtual void turn_command(double dir) = 0;
 };
 
 #endif //GAME_CLIENT_NETWORKCLIENT_H
